User Stories 7 + 21 (Spiel anlegen + laden)

+++ src/main/java/de/dhbw/karlsruhe/model/JPA/Rolle.java
@@ -12,8 +12,8 @@ public class Rolle {
     private long id;
     private String name;
 
-    public static final int ROLLE_TEILNEHMER = 1;
-    public static final int ROLLE_SPIELLEITER = 2;
+    public static final long ROLLE_TEILNEHMER = 1;
+    public static final long ROLLE_SPIELLEITER = 2;
 
     public long getId() {
         return id;
